update api
update api

update api

fix auth.js
