--- api/src/resolvers/Place.js
@@ -1,7 +1,0 @@
-function tags (parent, args, context) {
-  return context.prisma.place({ id: parent.id }).tags()
-}
-
-module.exports = {
-  tags,
-}
