--- api/src/resolvers/User.js
@@ -1,7 +1,0 @@
-function company(parent, args, context) {
-  return context.prisma.user({ id: parent.id }).company()
-}
-
-module.exports = {
-  company,
-}
